Fix validacion p/ elegir local

+++ app/Http/Controllers/CheckoutController.php
@@ -7,6 +7,7 @@ use Gloudemans\Shoppingcart\Facades\Cart;
 use Auth;
 use App\Order;
 use App\OrderItems;
+use Validator;
 
 class CheckoutController extends Controller
 {
@@ -38,30 +39,41 @@ class CheckoutController extends Controller
      */
     public function store(Request $request)
     {   
-        // Guardar orden en BD
-        $order = Order::create([
-            'user_id' => auth()->user()->id,
-            'store_address' => $request->store_address,
-            'subtotal' => Cart::subtotal(),
-            'tax' => Cart::tax(),
-            'payment' => $request->payment,
-            'total' => Cart::total(),
-            'status' => 'Pendiente'
+        $validator = Validator::make($request->all(), [
+            'store_address' => 'required',
+            'payment' => 'required'
         ]);
 
-        // Guardar items del carrito en BD
-        foreach (Cart::content() as $item) {
-            OrderItems::create([
-                'order_id' => $order->id,
-                'name' => $item->model->name,
-                'price' => $item->model->price,
-                'quantity' => $item->qty
+        if ($validator->fails()) {
+            return redirect()->route('checkout.index')->withErrors($validator);
+        } else {
+            // Guardar orden en BD
+            $order = Order::create([
+                'user_id' => auth()->user()->id,
+                'store_address' => $request->store_address,
+                'subtotal' => Cart::subtotal(),
+                'tax' => Cart::tax(),
+                'payment' => $request->payment,
+                'total' => Cart::total(),
+                'status' => 'Pendiente'
             ]);
+
+            // Guardar items del carrito en BD
+            foreach (Cart::content() as $item) {
+                OrderItems::create([
+                    'order_id' => $order->id,
+                    'name' => $item->model->name,
+                    'price' => $item->model->price,
+                    'quantity' => $item->qty
+                ]);
+            }
+
+            Cart::destroy();
+
+            return redirect('success');
         }
 
-        Cart::destroy();
-
-        return redirect('success');
+        
     }
 
     /**
